refactor: delete duplicate searchUser function
