+++ src/apis/WorkTime/index.ts
@@ -1,33 +1,15 @@
-import {
-  GoogleAuthProvider,
-  browserSessionPersistence,
-  setPersistence,
-  signInWithPopup,
-} from 'firebase/auth';
-import { auth, db } from '../firebase';
+import { db } from '../firebase';
 import {
   addDoc,
   collection,
   getDocs,
-  orderBy,
-  query,
   serverTimestamp,
-  where,
 } from '@firebase/firestore';
 import { WORK_TIME_COLLECTION } from 'constants/collection';
 import { ERROR_ALERT } from 'constants/alert';
 import { IuserData } from 'types/user';
 import { getSessionUserData } from 'utils/user';
-
-export const login = async () => {
-  try {
-    const provider = new GoogleAuthProvider();
-    await setPersistence(auth, browserSessionPersistence);
-    await signInWithPopup(auth, provider);
-  } catch (error) {
-    alert(ERROR_ALERT);
-  }
-};
+import { searchUser } from 'apis/User';
 
 export const addWorkTimeData = (workTime: number) => {
   const user: IuserData | undefined = getSessionUserData();
@@ -50,19 +32,9 @@ export interface IworkTimeResponse {
   name: string;
 }
 
-const searchUser = () => {
-  const user: IuserData | undefined = getSessionUserData();
-  const searchUserQuery = query(
-    collection(db, WORK_TIME_COLLECTION),
-    where('uid', '==', user?.uid),
-    orderBy('timeStamp', 'desc'),
-  );
-  return searchUserQuery;
-};
-
 export const getWorkTimeData = async () => {
   try {
-    const response = await getDocs(searchUser());
+    const response = await getDocs(searchUser(WORK_TIME_COLLECTION));
 
     const responseArray: IworkTimeResponse[] = [];
     response.forEach((doc) => {
